Outsource MapItems, using hotfix window.objects for globals
